Organized utils a little better so that Grid-only utils are in Grid/utils
