+++ source/Grid/utils/calculateSizeAndPositionDataAndUpdateScrollOffset.js
